Set up data repository from Chathran data
Following discussions with @BasFaggingerAuer I decided to split Chathran
into a mostly-C++ part which contains algorithms for generically
handling everything that needs handling in a game, and a mostly-Lua part
that defines what exactly appears in the game and how everything is
related.

In other words, Lua scripts define the structure of the game (the rules,
if you wish) while Chathran's C++ code defines the allowed `syntax' by
which rules can be constructed and the required evolution functions on
this syntax (i.e. it knows how the evolution of time changes the state
of Chathran). Mathematically one could think of Chathran as an object
initially constructed by Lua scripts on which Chathran's C++-code acts
as a time evolution operator.

At this stage I cannot tell whether it will be practical to keep this
distinction, however splitting Chathran's C++ algorithms from its assets
(with `assets' I mean data such as scripts and textures and stuff) seems
like a good way to help separate the specifics from the generics. I want
Chathran to be flexible to every possible extent, and that's easiest if
such flexibility is imposed by using an isolated set of data as input to
the program.

The program itself is thus not allowed to fix anything that can be left
variable, as long as such flexibility does not unduly make the codebase
complicated or slow. This is admittedly a vague statement but one must
draw a line somewhere between flexibility and practical coding. It will
be done on a case-by-case basis.