Adjusted start script for linking against Chathran
Since the data is now separate from Chathran proper, it needs to be
linked differently to be validly included as data. In this commit I
adjusted start.lua to properly set the data directory and to provide a
way to set the path (from inside Chathran) using Chathran cmake's
CHGAME_INCLUDE_DIR (which should be the root directory of this
repository).